The Quality Lab measures a render, and three of its own instruments were wrong
Phases 2 through 5 of docs/quality-lab/: the capture layer, the CLI, the
versioned report, the temporal detectors, the AOV-gated masks, and the first two
benchmark scenes. `avgen_quality analyze | compare | validate --ladder | version`,
a native C++ tool linking `avgen_core`, adding nothing to the engine's
dependencies and included by nothing in `src/`.

The instrument now has a motion-compensated residual: it reads the engine's own
velocity AOV, warps the previous frame by it, and differences only where an
identifier and a depth test agree the pixel has a history. That is the detector
ADR-243 did not have. Its control arm is the finding: a grating translating
exactly one pixel per frame -- pure authored motion, no defect in it at all --
gives a residual of 0.000 and a temporal alternation of 31.3. ADR-243's whole
mechanism, in a unit test, in under a second.

On `examples/quality/aliasing-dolly` the spatial measure reproduces ADR-243's
ordering with a new instrument on new content: FXAA off is +8.6% worse and
supersample 2x is -18.2% better, which is metrics.md 4.2's second gate. The
temporal measure agrees here, which it did not on Glowmere -- so ADR-243's
finding is narrowed to the axis it named as unexamined, and not overturned,
because nobody has looked at these frames.

ffmpeg 9.0.1 with libvmaf arrived, and Phase 2 gained VMAF, PSNR-HVS and CAMBI
without a design change -- the degradation contract doing what it is for. What
they cost was a day on the ladder, and three instruments came back wrong:

  * the default VMAF model scores a posterised gradient 100.0 against an
    identical pair's 97.3 (ADM reads added contrast as added detail). `neg` is
    now the default. Even under `neg` it ranks dither -- the remedy, which takes
    CAMBI from 22.99 to 0.00 -- below the banding.
  * CAMBI detects subtle banding and is blind to coarse posterisation, so
    metrics.md 4.1's banding arm had the direction backwards.
  * the native quantisationSteps proxy is blind the other way AND scores dither
    as banding: 0.000 banded, 0.031 dithered. Kept, scoped, and both failures
    asserted as ladder arms so they cannot drift into a trusted zero.

Both benchmark scenes were vacuous on the first attempt and both were caught by
the sequence-hash rule before a single metric ran. A spline camera takes its
target from further along its own path, so a lateral dolly aimed it at empty
background and rendered rms_contrast 0.0000. And `post/output/antialias`
defaults to 0, so `--disable fxaa` logged a DIAGNOSTIC render and wrote a
byte-identical hash -- ADR-182's exact shape, one layer up from where ADR-182
found it.

The ladder itself has a control now: run against a deliberately broken metric it
must fail, and the test asserts the targeted arms fail while the identity arm
still passes. Without that, "the ladder passes" says nothing.

Validated: 43 ladder checks plus the meta-control (`validate --ladder`, exit 0);
unit 2040 cases / 1,938,273 assertions, 4 skipped; render 303 / 410,310, 1
skipped. Renders under tools/gpu-lock.sh; no timing is claimed -- the load
average was 12-26 throughout and ADR-170 says structural quantities survive
contention and milliseconds do not.

Not validated: no human has looked at anything this instrument has measured.
Phase 6 is the gate, and every direction check so far is one instrument agreeing
with another.

ADR-252, ADR-253, ADR-254. metrics.md, artifact-detection.md,
benchmark-scenes.md and architecture.md are amended where implementation
corrected them.

+++ tools/quality-lab/metrics/spatial.cpp
@@ -76,8 +76,8 @@ double ssimPlanes(const std::vector<float>& a, const std::vector<float>& b, std:
             double sab = 0.0;
             for (std::uint32_t j = 0; j < kWin; ++j) {
                 for (std::uint32_t i = 0; i < kWin; ++i) {
-                    const double va = a[static_cast<std::size_t>(y + j) * w + x + i];
-                    const double vb = b[static_cast<std::size_t>(y + j) * w + x + i];
+                    const double va = static_cast<double>(a[static_cast<std::size_t>(y + j) * w + x + i]);
+                    const double vb = static_cast<double>(b[static_cast<std::size_t>(y + j) * w + x + i]);
                     sa += va;
                     sb += vb;
                     saa += va * va;
@@ -266,8 +266,11 @@ double spatialLaplacian(const Frame& frame) {
     for (std::uint32_t y = 1; y + 1 < frame.height; ++y) {
         for (std::uint32_t x = 1; x + 1 < w; ++x) {
             const std::size_t i = static_cast<std::size_t>(y) * w + x;
-            const double lap = 4.0 * plane[i] - plane[i - 1] - plane[i + 1] - plane[i - w] -
-                               plane[i + w];
+            const double lap = 4.0 * static_cast<double>(plane[i]) -
+                               static_cast<double>(plane[i - 1]) -
+                               static_cast<double>(plane[i + 1]) -
+                               static_cast<double>(plane[i - w]) -
+                               static_cast<double>(plane[i + w]);
             sum += std::fabs(lap);
             ++n;
         }
@@ -290,10 +293,10 @@ double quantisationSteps(const Frame& frame) {
             // neighbourhood is otherwise flat -- a gradient that should have been smooth arriving in
             // stairs. Texture fails the flatness test and dither fails the step test, which is what
             // keeps this from simply counting edges.
-            const double left = plane[i - 1];
-            const double right = plane[i + 1];
+            const double left = static_cast<double>(plane[i - 1]);
+            const double right = static_cast<double>(plane[i + 1]);
             const double step = std::fabs(right - left);
-            const double curvature = std::fabs(2.0 * plane[i] - left - right);
+            const double curvature = std::fabs(2.0 * static_cast<double>(plane[i]) - left - right);
             if (step > 3.0 && step < 12.0 && curvature > 1.5) {
                 ++stepped;
             }
